+++ static/elements/app-router/changelog.md
@@ -1,5 +1,37 @@
 ## app-router change log
 
+#### v2.3.2
+- Fixed bug where calling `router.go('/path')` on the current path wouldn't reload the page.
+- Switched `router.go('/path')` to fire a `popstate` event instead of directly calling `stateChange()` in order to support multiple routers on the same page.
+
+#### v2.3.1
+- Fixing bug where `router.go('/path')` would replace state instead of push state.
+
+#### v2.3.0
+- Adding `typecast="auto|string"` option on the `app-router`. Path variables and query parameters are typecast to numbers, booleans, and unescaped strings by default. Now you can get the raw string with `typecast="string"`.
+- Optimized hash fragment changes so that if only the hash fragment changes it will scroll to the fragment and not reload the entire page.
+
+#### v2.2.1
+- Fixing bug where the `before-data-binding` event wasn't using the updated model if the entire model was replaced.
+
+#### v2.2.0
+- Added ability to scroll to hash fragment on navigation. For example, `http://example.com/#/page1#middle` will now scroll to an element with `id="middle"` or `name="middle"`.
+
+#### v2.1.0
+- Added data binding to `<template>` tags when Polymer (`TemplateBinding.js`) is present.
+- Added `bindRouter` attribute to pass the router to the `app-route`'s page.
+- Added `before-data-binding` event to add properties to a model before it's bound to the route's custom element or template.
+- Fixed a `core-animated-pages` bug where multiple URLs matched the same `app-route` (ex: `path="/page/:num"` and paths `/page/1`, `/page/2`).
+
+#### v2.0.4
+- The move from `platform.js` to `webcomponents.js` removed the `URL()` constructor polyfill. The v2.0.3 fix created a bug in Safari when parsing the URL. This fixes Safari.
+
+#### v2.0.3
+- The move from `platform.js` to `webcomponents.js` removed the `URL()` constructor polyfill https://github.com/Polymer/webcomponentsjs/issues/53. IE doesn't support the `URL()` constructor yet so this fix is adding URL parse support for IE.
+
+#### v2.0.2
+- Fixing [issue 19](https://github.com/erikringsmuth/app-router/issues/19) using best effort approach. Use `template.createInstance()` if Polymer is loaded, otherwise use `document.importNode()`.
+
 #### v2.0.1
 - Fixing bug where multiple `<app-route>`s had an `active` attribute.